requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-MySQL==1.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
req==1.0.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
